requirements.txt
flask
lxml[html_clean]
newspaper4k
openai
weaviate-client